ee/api/requirements-crons.txt
# Keep this version to not have conflicts between requests and boto3
urllib3==1.26.16
requests==2.31.0
boto3==1.28.79
pyjwt==2.8.0
psycopg2-binary==2.9.9
elasticsearch==8.11.0
jira==3.5.2



fastapi==0.104.1
python-decouple==3.8
pydantic[email]==2.3.0
apscheduler==3.10.4

clickhouse-driver[lz4]==0.2.6
redis==5.0.1
azure-storage-blob==12.19.0
